Remove AbstractMetaFunction::invalid()
Change AbstractMetaBuilderPrivate::traverseFunction() to consistently
return 0 if something does not fit. Previously, functions with
half-parsed arguments with the invalid flag set were returned. This
caused strange side effects since the flag was not checked in
all places.

The only relevant information is whether some constructors were
rejected. In that case, no default constructors or default copy
constructors should be generated. This is now determined
by checking the code model function item; the attribute
HasRejectConstructor is introduced for this.

This fixes:

- Make it possible to reject the QTextStreamManipulator
  constructor taking a function pointer without having
  a default constructor generated:

  typedef void (QTextStream::*QTSMFI)(int);
  QTextStreamManipulator(QTSMFI m, int a) Q_DECL_NOTHROW

- Implement QtCharts whose class QAbstractSeries has a rejected
  constructor.

Change-Id: I6310574ba677dac20699f257340d2c2a55674353
Reviewed-by: Christian Tismer <tismer@stackless.com>