libshiboken: Add Array converters
Add a SbkArrayConverter struct which provides a list of
check functions that return a converter function
for an array of matching size.

Add simple array converters for arrays of C++ primitive types.

Instances of the ArrayHandle<>,  Array2Handle<> templates will be
generated which may point to internal data or allocated arrays.

Task-number: PYSIDE-354
Task-number: PYSIDE-516
Change-Id: I157606891fad345ccd7af6d4a9d4dcb0c634b2f4
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>