Fix building on openSuSE and Ubuntu 16.04 CI machines
Previously all the GCC header paths were explicitly passed to
libclang, which caused redefinition errors of builtin functions or
intrisics (like __rdtsc). Instead of passing the include paths
explicitly, we rely on libclang itself recognizing that there are
GCC paths in the default search locations.

Also we need to pass the libclang builtin headers location, because
it is not able to find them by itself. Usually the search location
for these headers is to get the executable path location (aka
the clang++ binary) and navigate to ../lib/clang/VERSION/include
relative to that binary. But because the shared library is used
instead of the binary, we need to explicitly pass that header
location via the -isystem flag.

Task-number: PYSIDE-513
Change-Id: I7c1127d85c0cea4c063c5c2a3548a1eef5eadaf3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>