Implement proper void pointer (void*) support
This change introduces a new type into the shiboken2 module which is
imported by calling "import PySide2.support.VoidPtr".

The type takes care of conversions from / to void* values in function
signatures.

Creating an instance can be done by passing either a shiboken wrapped
object, or an integer representing an address, or a python object that
implements the buffer interface.

For example, this is useful for passing numpy arrays to C OpenGL
functions that take void* parameters. First you convert the array into
a bytestring (using numpy.array.tobytes(), then you instantiate a
VoidPtr from that bytestring, and finally you pass it along to a GL
function.

One corner case that is currently not supported is void** parameters.

Change-Id: I01e291d6856cb6bd8b5175adc3ead6b728036535
Reviewed-by: Christian Tismer <tismer@stackless.com>