Typesystem parser: Look up external injected code in typesystem path
Make TypeDatabase::modifiedTypesystemFilepath() public and
add parameter preventing it from stripping the path so that it
can handle 'glue/snippet.cpp' correctly.

This removes the need to run shiboken with the working directory set to the
type system file path in order to find the external injected code snippets
for documentation generation.

Task-number: PYSIDE-363
Change-Id: Ic61cc1d1187e326e5255b29759a9d42e27ebd8d4
Reviewed-by: Christian Tismer <tismer@stackless.com>