Do not link shared libraries against the python shared library on Unix
There is no need to link against the python shared library (or static
library) on Linux / macOS because the python symbols will already be
loaded by the dynamic linker, by the time when a PySide2 module is
imported.

This is especially relevant on Debian distributions, which ship a
statically built python interpreter, as well as a python shared
library (which is not used by the interpreter). The python shared
library package is intended for native application that want to
embed a Python interpreter inside the application.

So far PySide2 would have linked against the python shared library,
which means that on Debian distributions when running a PySide2
example, the executable would load two versions of the same python
symbols (one from the static interpreter, and one from the shared
library), and this might cause issues.

On Windows we still need to link against the python.lib import library,
because otherwise the python symbols would not be resolved.

Task-number: PYSIDE-522
Change-Id: I4483ded96461c04f8e7a17d39068d07f772a973c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>