Fix qApp macro refcount
There are two borrowed references in the code,
so we need to manually increase the refcount.

Usually the PyEval_GetBuiltins and PyModule_GetDict
functions are used locally, so there is no real need
of taking care of the refcounts, but since we are using
it globally, and adjusting the refcount by ourselves,
it was necessary to add the missing references by hand.

Task-number: PYSIDE-585
Change-Id: Icc1e7719a6b5d3654d12ab37cd509a096821d7a6
Reviewed-by: Christian Tismer <tismer@stackless.com>