ShibokenGenerator::buildAbstractMetaTypeFromString(): Handle QFlags
The function would fail for a QFlags type defined in a dependent
typesystem file with errors like:

Can't write the C++ to Python conversion function for container type
'QPair': Could not find type '::QFlags<QAccessible::RelationFlag>' for
use in 'toPython' conversion.

for example caused by QAccessibleWidget::relation() where the flag is
defined in class QAccessible in QtGui.

The underlying reason is that there is no type entry for the class
QFlags. To work around this, detect it by looking at the instantiated
types and create a flags type accordingly.

Task-number: PYSIDE-487
Change-Id: Ib3c962aeedf8deb1a6b11976bd91316d39738310
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>