Limited API: remove the "slot_" workaround
It has been a problem to compile the file "typespec.*" when using
Python 2. For some reason, the name "slot" was not usable.

I cannot find out what it was and how it vanished, but now I could
remove the work-around without problems. Maybe it is due to some
different includes which don't define some macros?

Change-Id: If316c9c78a18cffcc5e4cac07610bd7f522db446
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>