shiboken: Improve handling of non-type template integer parameters
When trying to specify the base class
template<int R, int C> QGenericMatrix of the QMatrixRxN classes,
clashes of the SBK..IDX enumeration values occurred since
the integers were not part of the type name.

Fix that by dynamically adding dummy entries of
EnumValueTypeEntry for the integer values encountered on the fly.

Task-number: PYSIDE-725
Change-Id: Ie6b4489b5293e04b7c2c76861341c99b136cd558
Reviewed-by: Christian Tismer <tismer@stackless.com>